refactor: move cheatsheet to global components
