+++ frontend/src/components/cheatsheet/cheatsheet-extension.ts
