+++ frontend/src/components/cheatsheet/cheatsheet.module.scss
