+++ frontend/src/components/editor-page/app-bar/app-bar.tsx
@@ -1,16 +1,16 @@
 /*
- * SPDX-FileCopyrightText: 2022 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { useApplicationState } from '../../../hooks/common/use-application-state'
 import { useOutlineButtonVariant } from '../../../hooks/dark-mode/use-outline-button-variant'
+import { CheatsheetButton } from '../../cheatsheet/cheatsheet-button'
 import { NewNoteButton } from '../../common/new-note-button/new-note-button'
 import { ShowIf } from '../../common/show-if/show-if'
 import { SignInButton } from '../../landing-layout/navigation/sign-in-button'
 import { UserDropdown } from '../../landing-layout/navigation/user-dropdown'
 import { SettingsButton } from '../../layout/settings-dialog/settings-button'
-import { CheatsheetButton } from './cheatsheet/cheatsheet-button'
 import { HelpButton } from './help-button/help-button'
 import { NavbarBranding } from './navbar-branding'
 import { ReadOnlyModeButton } from './read-only-mode-button'
