+++ frontend/src/extensions/essential-app-extensions/task-list/task-list-checkbox-app-extension.ts
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import type { CheatsheetExtension } from '../../../components/editor-page/cheatsheet/cheatsheet-extension'
+import type { CheatsheetExtension } from '../../../components/cheatsheet/cheatsheet-extension'
 import type { MarkdownRendererExtensionOptions } from '../../_base-classes/app-extension'
 import { AppExtension } from '../../_base-classes/app-extension'
 import { SetCheckboxInCheatsheet } from './set-checkbox-in-cheatsheet'
