+++ frontend/src/extensions/external-lib-app-extensions/plantuml/plantuml-app-extension.ts
@@ -3,7 +3,7 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import type { CheatsheetExtension } from '../../../components/editor-page/cheatsheet/cheatsheet-extension'
+import type { CheatsheetExtension } from '../../../components/cheatsheet/cheatsheet-extension'
 import {
   basicCompletion,
   codeFenceRegex
